requirements.txt
flask==1.1.1
firebase-admin==3.2.1
google-cloud-pubsub==1.1.0
gunicorn==20.0.4
psaw==0.0.7
requests==2.22.0
simplejson==3.17.0
